Refactor into package with main entry
