--- app.py
@@ -1,4 +1,0 @@
-from gui import main
-
-if __name__ == '__main__':
-    main()
