+++ main.py
@@ -1,0 +1,4 @@
+from rpgwiki.__main__ import run
+
+if __name__ == '__main__':
+    run()
