+++ rpgwiki/__main__.py
@@ -1,0 +1,7 @@
+from .gui import main
+
+def run() -> None:
+    main()
+
+if __name__ == '__main__':
+    run()
